+++ CardPayments/src/test/java/com/paypal/android/cardpayments/CardClientUnitTest.kt
@@ -26,6 +26,7 @@ import kotlinx.coroutines.test.advanceUntilIdle
 import kotlinx.coroutines.test.resetMain
 import kotlinx.coroutines.test.runTest
 import kotlinx.coroutines.test.setMain
+import kotlinx.serialization.InternalSerializationApi
 import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
@@ -40,6 +41,7 @@ private val AUTH_CHALLENGE_URL = Uri.parse("https://fake.com/url")
 
 @ExperimentalCoroutinesApi
 @RunWith(RobolectricTestRunner::class)
+@OptIn(InternalSerializationApi::class)
 class CardClientUnitTest {
 
     private val card = Card("4111111111111111", "01", "24", "123")