+++ Demo/src/main/java/com/paypal/android/models/OrderRequest.kt
@@ -1,7 +1,9 @@
 package com.paypal.android.models
 
 import com.paypal.android.api.model.OrderIntent
+import kotlinx.serialization.Serializable
 
+@Serializable
 data class OrderRequest(
     val intent: OrderIntent,
     val shouldVault: Boolean = false