+++ Demo/src/main/java/com/paypal/android/usecase/CreatePayPalSetupTokenUseCase.kt
@@ -1,8 +1,10 @@
 package com.paypal.android.usecase
 
-import com.google.gson.JsonObject
-import com.google.gson.JsonParser
 import com.paypal.android.api.model.PayPalSetupToken
+import com.paypal.android.api.model.serialization.PayPalDetails
+import com.paypal.android.api.model.serialization.PayPalExperienceContext
+import com.paypal.android.api.model.serialization.PayPalSetupRequestBody
+import com.paypal.android.api.model.serialization.PayPalSource
 import com.paypal.android.api.services.SDKSampleServerAPI
 import com.paypal.android.api.services.SDKSampleServerResult
 import kotlinx.coroutines.Dispatchers
@@ -15,26 +17,19 @@ class CreatePayPalSetupTokenUseCase @Inject constructor(
 
     suspend operator fun invoke(): SDKSampleServerResult<PayPalSetupToken, Exception> =
         withContext(Dispatchers.IO) {
-            // language=JSON
-            val request = """
-            {
-              "paymentSource": {
-                "paypal": {
-                  "usageType": "MERCHANT",
-                  "experienceContext": {
-                    "vaultInstruction": "ON_PAYER_APPROVAL",
-                    "returnUrl": "com.paypal.android.demo://vault/success",
-                    "cancelUrl": "com.paypal.android.demo://vault/cancel"
-                  }
-                }
-              }
-            }
-        """
+            val payPalSetupRequest = PayPalSetupRequestBody(
+                paymentSource = PayPalSource(
+                    paypal = PayPalDetails(
+                        usageType = "MERCHANT",
+                        experienceContext = PayPalExperienceContext(
+                            vaultInstruction = "ON_PAYER_APPROVAL",
+                            returnUrl = "com.paypal.android.demo://vault/success",
+                            cancelUrl = "com.paypal.android.demo://vault/cancel"
+                        )
+                    )
+                )
+            )
 
-            // Ref: https://stackoverflow.com/a/19610814
-            val body = request.replace("\\/", "/")
-
-            val jsonOrder = JsonParser.parseString(body) as JsonObject
-            sdkSampleServerAPI.createPayPalSetupToken(jsonOrder)
+            sdkSampleServerAPI.createPayPalSetupToken(payPalSetupRequest)
         }
 }